feat: Phase D — BasicHmm<Obs> template HMM core
- Add include/libhmm/basic_hmm.h:
  template<typename Obs = double> class BasicHmm
  - emis_ typed as vector<unique_ptr<BasicEmissionDistribution<Obs>>>
  - initializeMatrices() creates default GaussianDistribution only for
    Obs = double (if constexpr); non-scalar HMMs require explicit
    setDistribution() calls
  - setDistribution() / getDistribution() use BasicEmissionDistribution<Obs>
  - All constructors, validate(), setTrans(), setPi() preserved unchanged
  - Non-copyable, movable; virtual destructor for inheritance
  - friend operator>> removed (uses only public interface)

- Rewrite include/libhmm/hmm.h as v3 compatibility header:
  using Hmm = BasicHmm<double>
  Declares operator<< and operator>> for Hmm
  src/hmm.cpp needs zero changes

Scalar path is zero-overhead: BasicHmm<double> identical to v3 Hmm.
All downstream code compiles unchanged through the alias. 41/41 tests pass.
